+++ packages/kit/src/runtime/client/types.d.ts
@@ -4,6 +4,7 @@ import {
 	CSRPageNode,
 	CSRPageNodeLoader,
 	CSRRoute,
+	CSRRouteServer,
 	ServerDataNode,
 	TrailingSlash,
 	Uses
@@ -12,13 +13,18 @@ import { Page, ParamMatcher } from '@sveltejs/kit';
 
 export interface SvelteKitApp {
 	/**
-	 * A list of all the error/layout/page nodes used in the app
+	 * A list of all the error/layout/page nodes used in the app.
+	 * - In case of router.resolution=client, this is filled completely upfront.
+	 * - In case of router.resolution=server, this is filled with the root layout and root error page
+	 *   at the beginning and then filled up as the user navigates around the app, loading new nodes
 	 */
 	nodes: CSRPageNodeLoader[];
 
 	/**
 	 * A list of all layout node ids that have a server load function.
 	 * Pages are not present because it's shorter to encode it on the leaf itself.
+	 *
+	 * In case of router.resolution=server, this only contains one entry for the root layout.
 	 */
 	server_loads: number[];
 
@@ -27,9 +33,16 @@ export interface SvelteKitApp {
 	 * is parsed into an array of routes on startup. The numbers refer to the indices in `nodes`.
 	 * If the leaf number is negative, it means it does use a server load function and the complement is the node index.
 	 * The route layout and error nodes are not referenced, they are always number 0 and 1 and always apply.
+	 *
+	 * In case of router.resolution=server, this object is empty, as resolution happens on the server.
 	 */
 	dictionary: Record<string, [leaf: number, layouts: number[], errors?: number[]]>;
 
+	/**
+	 * A map of `[matcherName: string]: (..) => boolean`, which is used to match route parameters.
+	 *
+	 * In case of router.resolution=server, this object is empty, as resolution happens on the server.
+	 */
 	matchers: Record<string, ParamMatcher>;
 
 	hooks: ClientHooks;
@@ -108,6 +121,8 @@ export interface HydrateOptions {
 	node_ids: number[];
 	params: Record<string, string>;
 	route: { id: string | null };
+	/** Only used when `router.resolution=server`; can then still be undefined in case of 404 */
+	server_route?: CSRRouteServer;
 	data: Array<ServerDataNode | null>;
 	form: Record<string, any> | null;
 }