+++ packages/kit/src/runtime/server/page/index.js
@@ -1,7 +1,7 @@
 import { text } from '../../../exports/index.js';
 import { compact } from '../../../utils/array.js';
 import { get_status, normalize_error } from '../../../utils/error.js';
-import { add_data_suffix } from '../../../utils/url.js';
+import { add_data_suffix } from '../../pathname.js';
 import { Redirect } from '../../control.js';
 import { redirect_response, static_error_page, handle_error_and_jsonify } from '../utils.js';
 import {