+++ packages/kit/test/apps/basics/playwright.config.js
@@ -5,8 +5,8 @@ export default {
 	...config,
 	webServer: {
 		command: process.env.DEV
-			? 'cross-env PUBLIC_PRERENDERING=false pnpm dev'
-			: 'cross-env PUBLIC_PRERENDERING=true pnpm build && pnpm preview',
+			? `cross-env PUBLIC_PRERENDERING=false ROUTER_RESOLUTION=${process.env.ROUTER_RESOLUTION ?? 'client'} pnpm dev`
+			: `cross-env PUBLIC_PRERENDERING=true ROUTER_RESOLUTION=${process.env.ROUTER_RESOLUTION ?? 'client'} pnpm build && pnpm preview`,
 		port: process.env.DEV ? 5173 : 4173
 	}
 };